requirements.txt
streamlit==1.28.1
groq==0.32.0
pandas==2.1.3
plotly==5.17.0
python-dotenv==1.0.0
